requirements.txt
click==7.1.2
azure-identity==1.5.0
azure-keyvault-secrets==4.2.0